# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 临床中心台 SiteDesk API
  version: 0.2.0
  description: |-
    **本文件是产物。** 契约的唯一定义源是 `packages/contracts/src` 下的 zod schema。

    分层：L1 资源层（CRUD 与列表）· L2 命令层（领域动作，返回 sideEffects）· L3 投影层（只读聚合，永不返回明细）。

    带 `x-gated-by` 的字段受列维度权限管辖：**无权限时该字段从响应中消失，不是返回 null**。

    行范围之外的资源一律返回 404 而非 403 —— 403 等于确认「它存在，只是你不能碰」，这个确认本身就是泄漏。
servers:
  - url: https://api.sitedesk.example
    description: 生产
...
paths:
...
  /v1/handovers:
    get:
      operationId: listHandovers
      tags:
        - site
      summary: 交接列表
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: status
          in: query
          required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/HandoverStatus'
      responses:
        '200':
          description: 交接列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListHandoversResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
    post:
      operationId: createHandover
      tags:
        - site
      summary: 发起交接
      description: |-
        休假、离职、调岗 —— 中心不会因此停下。
        只能交接自己当前负责的中心；接手人必须是同工种的在职人员。
      x-layer: L1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHandoverRequest'
      responses:
        '201':
          description: 发起交接
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Handover'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/handovers/{id}:complete:
    post:
      operationId: completeHandover
      tags:
        - site
      summary: 确认交接完成
      description: |-
        清单未逐项确认不得完成 —— 交接单签了字但受试者没交底，等于没交接。
        完成时把派工从原负责人转到接手人，两边的行范围随即改变。
      x-layer: L2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteHandoverRequest'
      responses:
        '200':
          description: 确认交接完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteHandoverResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 幂等键被不同的请求体复用
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/idempotency-key-reused
                title: 幂等键被不同的请求体复用
                status: 409
                code: idempotency-key-reused
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/handovers/{id}/items/{seq}:done:
    post:
      operationId: completeHandoverItem
      tags:
        - site
      summary: 确认交接清单的某一项
      description: 逐项确认，不是一次打勾了事 —— 交接单签了字但受试者没交底，等于没交接。
      x-layer: L2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: seq
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 9007199254740991
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteHandoverItemRequest'
      responses:
        '200':
          description: 确认交接清单的某一项
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteHandoverItemResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 幂等键被不同的请求体复用
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/idempotency-key-reused
                title: 幂等键被不同的请求体复用
                status: 409
                code: idempotency-key-reused
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
  /v1/staff:
    get:
      operationId: listStaff
      tags:
        - site
      summary: 人员与派工
      description: 外部方看不到员工名册 —— 那与机构履行监管职责无关。
      x-layer: L1
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: limit
          in: query
          required: false
          schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
        - name: roleKind
          in: query
          required: false
          schema:
            allOf:
              - $ref: '#/components/schemas/RoleKind'
        - name: successionGap
          in: query
          required: false
          description: 只看「带多个中心却无继任者」的人
          schema:
            type: boolean
      responses:
        '200':
          description: 人员与派工
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStaffResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/startup-items/{id}:complete:
    post:
      operationId: completeStartupItem
      tags:
        - site
      summary: 标记启动清单项完成
      description: |-
        完成必须同时记下时间与人 —— 只记「做完了」而不记「谁做的」，核查时说不清。
        若本次完成清空了最后一个阻塞项，`sideEffects` 会明确告知「可推进 SIV」。
      x-layer: L2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteStartupItemRequest'
      responses:
        '200':
          description: 标记启动清单项完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteStartupItemResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 版本冲突：资源已被他人修改
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/conflict-version
                title: 版本冲突：资源已被他人修改
                status: 409
                code: conflict-version
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
  /v1/startup-items/{id}:reopen:
    post:
      operationId: reopenStartupItem
      tags:
        - site
      summary: 撤销启动清单项的完成标记
      description: 撤销是敏感动作：它可能让一个已经推进的中心回到「其实没准备好」的状态，必须写原因。
      x-layer: L2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        - name: Idempotency-Key
          in: header
          required: true
          description: 幂等键（uuid）。24 小时内重放同一键返回首次结果。
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisableAccountRequest'
      responses:
        '200':
          description: 撤销启动清单项的完成标记
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReopenStartupItemResponse'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '409':
          description: 版本冲突：资源已被他人修改
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/conflict-version
                title: 版本冲突：资源已被他人修改
                status: 409
                code: conflict-version
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
  /v1/study-sites/{id}/startup-items:
    get:
      operationId: getStartupChecklist
      tags:
        - site
      summary: 中心启动清单
      description: |-
        8 类清单项，带负责人、应完成日与**阻塞项**标记。
        阻塞项未清零，`POST /v1/study-sites/{id}:advance` 到 `siv` 会被闸门拦下。
      x-layer: L1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      responses:
        '200':
          description: 中心启动清单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartupChecklist'
        '401':
          description: 未认证
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/unauthenticated
                title: 未认证
                status: 401
                code: unauthenticated
        '403':
          description: 无此动作权限
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/forbidden-action
                title: 无此动作权限
                status: 403
                code: forbidden-action
        '404':
          description: 资源不存在
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/not-found
                title: 资源不存在
                status: 404
                code: not-found
        '422':
          description: 请求校验未通过
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/validation-failed
                title: 请求校验未通过
                status: 422
                code: validation-failed
        '429':
          description: 请求过于频繁
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/rate-limited
                title: 请求过于频繁
                status: 429
                code: rate-limited
        '500':
          description: 服务内部错误
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: https://sitedesk.dev/problems/internal
                title: 服务内部错误
                status: 500
                code: internal
...
components:
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
          description: https://sitedesk.dev/problems/{code}
        title:
          type: string
        status:
          type: integer
          minimum: -9007199254740991
          maximum: 9007199254740991
        code:
          type: string
          enum:
            - validation-failed
            - unauthenticated
            - forbidden-action
            - not-found
            - conflict-version
            - idempotency-key-reused
            - invariant-violated
            - gate-not-satisfied
            - rate-limited
            - internal
        detail:
          type: string
        instance:
          description: 出错的请求路径
          type: string
        traceId:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/FieldIssue'
        unmet:
          type: array
          items:
            $ref: '#/components/schemas/GateUnmet'
        invariant:
          type: string
      required:
        - type
        - title
        - status
        - code
      additionalProperties: false
      description: RFC 9457 Problem Details。所有非 2xx 响应都是这个形状。
...
    RoleKind:
      type: string
      enum:
        - CRA
        - CRC
        - PM
        - QA
        - DM
...
    HandoverStatus:
      type: string
      enum:
        - pending
        - completed
        - cancelled
...
    DisableAccountRequest:
      type: object
      properties:
        reason:
          type: string
          minLength: 4
          maxLength: 500
          description: 变更原因。会写入审计轨迹；核查时真正被问的就是这一栏。
      required:
        - reason
      additionalProperties: false
...
    ListHandoversResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Handover'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
    CreateHandoverRequest:
      type: object
      properties:
        toAccountId:
          type: string
          format: uuid
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        studySiteIds:
          minItems: 1
          type: array
          items:
            type: string
            format: uuid
            pattern: >-
              ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
        reason:
          type: string
          minLength: 5
          maxLength: 500
        plannedOn:
          type: string
          format: date
          pattern: >-
            ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
      required:
        - toAccountId
        - studySiteIds
        - reason
        - plannedOn
      additionalProperties: false
    CompleteHandoverRequest:
      type: object
      properties: {}
      additionalProperties: false
    CompleteHandoverResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Handover'
        sideEffects:
          type: array
          items:
            $ref: '#/components/schemas/SideEffect'
          description: 本次命令连带触发的事件。为空数组表示没有连带影响。
      required:
        - data
        - sideEffects
      additionalProperties: false
    CompleteHandoverItemRequest:
      type: object
      properties: {}
      additionalProperties: false
    CompleteHandoverItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Handover'
        sideEffects:
          type: array
          items:
            $ref: '#/components/schemas/SideEffect'
          description: 本次命令连带触发的事件。为空数组表示没有连带影响。
      required:
        - data
        - sideEffects
      additionalProperties: false
...
    ListStaffResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Staff'
        nextCursor:
          nullable: true
          description: null 表示已到末页
          type: string
          minLength: 1
          maxLength: 512
      required:
        - items
        - nextCursor
      additionalProperties: false
    CompleteStartupItemRequest:
      type: object
      properties:
        note:
          type: string
          maxLength: 500
      additionalProperties: false
    CompleteStartupItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/StartupItem'
        sideEffects:
          type: array
          items:
            $ref: '#/components/schemas/SideEffect'
          description: 本次命令连带触发的事件。为空数组表示没有连带影响。
      required:
        - data
        - sideEffects
      additionalProperties: false
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://sso.example/.well-known/openid-configuration
      description: 内部员工：企业微信 / 飞书 OIDC
    magicLink:
      type: http
      scheme: bearer
      description: 外部方（机构办 / 研究者）：一次性魔法链接换取的短期会话令牌
security:
  - oidc: []
  - magicLink: []